requirements.txt
fastapi==0.116.1
Jinja2==3.1.6
pandas==2.3.1
pydantic==2.11.7
python_jose==3.5.0
Requests==2.32.4
SQLAlchemy==2.0.41
uvicorn==0.34.0
html5lib==1.1
psycopg2-binary==2.9.10
fastapi-users[sqlalchemy,oauth]==11.0.0
bcrypt==4.3.0
oauthlib==3.2.0
asyncpg
httpx-oauth
python-dotenv